requirements.txt
eth_account==0.13.7
python-dotenv==1.2.1
requests==2.32.5
selenium==4.39.0
toml==0.10.2
typer==0.21.1
web3==7.14.0
xrpl-py==4.4.0